relational_image_generation_evaluation/requirements.txt
datasets==2.12.0
ftfy==6.1.1
gdown==4.7.1
huggingface_hub==0.13.4
numpy==1.23.5
Pillow==9.3.0
Pillow==9.5.0
pytorch_lightning==2.0.1.post0
regex==2022.7.9
setuptools==65.6.3
timm==0.6.13
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.0
tqdm==4.65.0
transformers==4.29.2
wandb==0.14.2
